Don't blank out data if validation fails on first step
