+++ libs/newsletter-workflow/src/lib/steps/createNewsletterLayout.ts
@@ -1,8 +1,5 @@
-import type {
-	WizardFormData,
-	WizardStepLayout,
-} from '@newsletters-nx/state-machine';
-import { formSchemas } from '@newsletters-nx/state-machine';
+import type { WizardStepLayout } from '@newsletters-nx/state-machine';
+import { executeCreate } from '../executeCreate';
 
 export const createNewsletterLayout: WizardStepLayout = {
 	markdownToDisplay: `# Create a newsletter
@@ -29,38 +26,7 @@ For example:
 				const name = stepData.formData ? stepData.formData['name'] : undefined;
 				return name ? undefined : 'NO NAME PROVIDED';
 			},
-			executeStep: async (
-				stepData,
-				stepLayout,
-				storageInstance,
-			): Promise<WizardFormData | string> => {
-				const schema = formSchemas['createNewsletter'];
-				if (!storageInstance) {
-					throw new Error('no storageInstance');
-				}
-				const parseResult = schema.safeParse(stepData.formData);
-
-				if (!parseResult.success) {
-					return `Form data is invalid for schema: ${
-						schema.description ?? '[no description]'
-					}`;
-				}
-
-				const storageResponse = await storageInstance.createDraftNewsletter({
-					...parseResult.data,
-					listId: undefined,
-				});
-
-				if (storageResponse.ok) {
-					console.log(
-						'createNewsletter step has updated storage.',
-						storageInstance,
-					);
-					return storageResponse.data;
-				}
-
-				return storageResponse.message;
-			},
+			executeStep: executeCreate,
 		},
 	},
 };
